Move category_alignment into diagnose stage.
Fixes #2998

Co-locate the root-cause category alignment heuristic with its sole
consumer in the diagnose stage instead of keeping it at app/agent/.
